% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenopath.R
\name{interaction_effects}
\alias{interaction_effects}
\title{Get the interaction effect sizes}
\usage{
interaction_effects(phenopath_fit)
}
\arguments{
\item{phenopath_fit}{An object of class \code{phenopath_fit}}
}
\value{
TODO
}
\description{
Get the interaction effect sizes
}
\examples{
sim <- simulate_phenopath() # returns a list with gene expression in y and covariates in x
fit <- phenopath(sim$y, sim$x, elbo_tol = 1e-2)
beta <- interaction_effects(fit)
}

