% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenopath.R
\name{interaction_sds}
\alias{interaction_sds}
\title{Get the interaction standard deviations}
\usage{
interaction_sds(phenopath_fit)
}
\arguments{
\item{phenopath_fit}{An object of class \code{phenopath_fit}}
}
\value{
TODO
}
\description{
Get the interaction standard deviations
}
\examples{
sim <- simulate_phenopath() # returns a list with gene expression in y and covariates in x
fit <- phenopath(sim$y, sim$x, elbo_tol = 1e-2)
beta_sd <- interaction_sds(fit)
}
