% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_elbo}
\alias{plot_elbo}
\title{Plots the ELBO}
\usage{
plot_elbo(fit)
}
\arguments{
\item{fit}{An object returned by a call to \code{phenopath}}
}
\value{
A \code{ggplot2} object of the ELBO against the number of iterations
}
\description{
Plots the evidence lower bound (ELBO) as a function of iterations
}
\examples{
sim <- simulate_phenopath()
fit <- phenopath(sim$y, sim$x)
plot_elbo(fit)
}
