% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenopath.R
\name{print.phenopath_fit}
\alias{print.phenopath_fit}
\title{Print a PhenoPath fit}
\usage{
\method{print}{phenopath_fit}(x, ...)
}
\arguments{
\item{x}{A \code{phenopath_fit} returned by a call to \code{phenopath}}

\item{...}{Additional arguments}
}
\value{
A string representation of a \code{phenopath_fit} object.
}
\description{
Print a PhenoPath fit
}
\examples{
sim <- simulate_phenopath() 
fit <- phenopath(sim$y, sim$x, elbo_tol = 1e-2)
print(fit)
}
