% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{sample_fns}
\alias{sample_fns}
\alias{sample_de}
\alias{sample_fns}
\alias{sample_pst}
\alias{sample_fns}
\alias{sample_pst_beta}
\alias{sample_fns}
\alias{sample_de_pst_beta}
\title{Sample parameters for simulation}
\usage{
sample_de()

sample_pst()

sample_pst_beta()

sample_de_pst_beta()
}
\description{
Sample parameters from de regime

Sample parameters from pseudotime regime

Sample parameters from pseudotime-interaction regime

Sample paramters from de-pseudotime-interaction regime
}
\keyword{internal}
