% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{simulate_one_gene}
\alias{simulate_one_gene}
\title{Simulate one gene}
\usage{
simulate_one_gene(N, pst, x, alpha = 0, c = 0, beta = 0, tau = 1e+06)
}
\description{
Simulate one gene from the model given paramters, z and covariates
}
\keyword{internal}

