% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenopath.R
\name{phenopath}
\alias{phenopath}
\title{PhenoPath - genomic trajectories with heterogeneous backgrounds}
\usage{
phenopath(exprs_obj, x, ...)
}
\arguments{
\item{exprs_obj}{Input gene expression, either
\enumerate{
\item An \code{ExpressionSet} object such as an \code{SCESet}, \emph{or}
\item A cell-by-gene matrix of normalised expression values in log form.
}}

\item{x}{The covariate vector, either
\enumerate{
\item The name of a column of \code{pData(exprs_obj)} if \code{exprs_obj} is an
\code{ExpressionSet}, \emph{or}
\item A numeric of factor vector of length equal to the number of cells
}}
}
\description{
PhenoPath learns genomic trajectories in the presence of heterogenous environmental
and genetic backgrounds using scalable Bayesian variational inference.
}
\details{
If an \code{ExpressionSet} is provided, \code{exprs(...)} is used

If \code{x} is the name of...
}

