% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenopath.R
\name{trajectory}
\alias{trajectory}
\title{Get the latent PhenoPath trajectory}
\usage{
trajectory(phenopath_fit)
}
\arguments{
\item{phenopath_fit}{An object of class \code{phenopath_fit}}
}
\value{
A vector of latent trajectory (pseudotime) values
}
\description{
Get the latent PhenoPath trajectory
}
\examples{
sim <- simulate_phenopath() # returns a list with gene expression in y and covariates in x
fit <- phenopath(sim$y, sim$x, elbo_tol = 1e-2)
z <- trajectory(fit)

}
